Deepcopy
No matter what  I tried I couldn't get the a copy of a dictionary to remain unaltered. I had to learn about a "deepcopy" that allows me to make a copy and change it around without changing the original dictionary.

https://stackoverflow.com/questions/2465921/how-to-copy-a-dictionary-and-only-edit-the-copy

https://www.programiz.com/python-programming/methods/list/index

I also used the index() method on the call to change the dictionary, which was an attempt to fix the original problem, but it was futile. Im pretty sure the commented line beneath it will now produce the same affect, but I am using an additional method so Im keeping it for learning purpose.