Added balance_teams
The balance teams function is defined rather weird, I think it needs a loop instead of how I defined it to be able to achieve some of the extra tasks. It all works though so I am moving to finishing th eidplsy stats funcitonality, then will revise everything to implement the extra tasks.