Finished Functionality
The program is complete and functionaly to specification. There are somethings I don't like about it, like how there is no consolidated for loops, or it doesnt use f strings as much as i would like it to, and it also wont work if another team is added to constants.py. There are aalso alot of functions im not sure are enccessary or 'pythonic'. The directions for the project are also bothering me, namingly the repromting of the main menu. The sample says ' press enter to continue' so Im assuming if you dont press enter it will quit or raise an error and have you press enter, and raising an error would be really invalid because saying 'press enter to continue' is giving the option of pressing something else to quit. Also, the way clean_data function in the directions seems like its going to scramble and dissasociate the data, which I really wanted to keep it contained and together right up untill it needed to be accessed to minize potential error.